+++ src/sys/mem/ptr/reexports.rs
@@ -5,16 +5,6 @@
 
 use crate::code::reexport;
 
-// macros
-reexport! { rust: core::ptr,
-    doc: "Create a `const` raw pointer to a place, without creating an intermediate reference.",
-    @addr_of as addr_of
-}
-reexport! { rust: core::ptr,
-    doc: "Create a `mut` raw pointer to a place, without creating an intermediate reference.",
-    @addr_of_mut as addr_of_mut
-}
-
 // structs
 reexport! { rust: core::ptr,
     doc: "`*mut T` but non-zero and *covariant*.",